Fix 64-bit pointer truncation and signed stack values in bgdi

+++ core/bgdrtm/src/copy.c
@@ -202,18 +202,18 @@ static int copytype( void * dst, void * src, DCB_TYPEDEF * var )
  *  Copy struct data from src to dst, using the information varspace given
  **/
 
-int bgd_copy_struct( INSTANCE * my, int * params )
+int bgd_copy_struct( INSTANCE * my, intptr_t * params )
 {
     return ( int ) copytypes(( void * )params[0], ( void * )params[1], ( DCB_TYPEDEF * )params[2], params[3], params[4] );
 }
 
-int bgd_internal_memcopy( INSTANCE * my, int * params )
+int bgd_internal_memcopy( INSTANCE * my, intptr_t * params )
 {
     memmove(( void * )params[0], ( void * )params[1], params[2] ) ;
     return 1 ;
 }
 
-int bgd_internal_copy_string_array( INSTANCE * my, int * params )
+int bgd_internal_copy_string_array( INSTANCE * my, intptr_t * params )
 {
     int n = params[ 2 ];
     while( n-- )
