+++ modules/mod_mathi/mod_mathi.c
@@ -1,7 +1,7 @@
 /*
- *  Copyright © 2006-2013 SplinterGU (Fenix/Bennugd)
- *  Copyright © 2002-2006 Fenix Team (Fenix)
- *  Copyright © 1999-2002 José Luis Cebrián Pagüe (Fenix)
+ *  Copyright &#169; 2006-2013 SplinterGU (Fenix/Bennugd)
+ *  Copyright &#169; 2002-2006 Fenix Team (Fenix)
+ *  Copyright &#169; 1999-2002 José Luis Cebrián Pagüe (Fenix)
  *
  *  This file is part of Bennu - Game Development
  *
@@ -39,7 +39,7 @@
 /* Funciones matemáticas */
 /* --------------------------------------------------------------------------- */
 
-static int math_abs( INSTANCE * my, int * params )
+static int math_abs( INSTANCE * my, intptr_t * params )
 {
     float num = *( float * ) &params[0] ;
     float res = ( num < 0 ) ? -num : num ;
@@ -48,7 +48,7 @@ static int math_abs( INSTANCE * my, int * params )
 
 /* --------------------------------------------------------------------------- */
 
-static int math_pow( INSTANCE * my, int * params )
+static int math_pow( INSTANCE * my, intptr_t * params )
 {
     float res = ( float )pow( *( float * ) & params[0], *( float * ) & params[1] ) ;
     return *(( int * )&res ) ;
@@ -56,7 +56,7 @@ static int math_pow( INSTANCE * my, int * params )
 
 /* --------------------------------------------------------------------------- */
 
-static int math_sqrt( INSTANCE * my, int * params )
+static int math_sqrt( INSTANCE * my, intptr_t * params )
 {
     float res = ( float )sqrt( *( float * ) & params[0] ) ;
     return *(( int * )&res ) ;
@@ -64,56 +64,56 @@ static int math_sqrt( INSTANCE * my, int * params )
 
 /* --------------------------------------------------------------------------- */
 
-static int math_cos( INSTANCE * my, int * params )
+static int math_cos( INSTANCE * my, intptr_t * params )
 {
     return cos(( double )( params[0] * M_PI / 180000.0 )) * 1000 ;
 }
 
 /* --------------------------------------------------------------------------- */
 
-static int math_sin( INSTANCE * my, int * params )
+static int math_sin( INSTANCE * my, intptr_t * params )
 {
     return sin(( double )( params[0] * M_PI / 180000.0 )) * 1000 ;
 }
 
 /* --------------------------------------------------------------------------- */
 
-static int math_tan( INSTANCE * my, int * params )
+static int math_tan( INSTANCE * my, intptr_t * params )
 {
     return tan(( double )( params[0] * M_PI / 180000.0 )) * 1000 ;
 }
 
 /* --------------------------------------------------------------------------- */
 
-static int math_acos( INSTANCE * my, int * params )
+static int math_acos( INSTANCE * my, intptr_t * params )
 {
     return acos(( double )( params[0] * 180000.0 / M_PI )) * 1000 ;
 }
 
 /* --------------------------------------------------------------------------- */
 
-static int math_asin( INSTANCE * my, int * params )
+static int math_asin( INSTANCE * my, intptr_t * params )
 {
     return asin(( double )( params[0] * 180000.0 / M_PI )) * 1000 ;
 }
 
 /* --------------------------------------------------------------------------- */
 
-static int math_atan( INSTANCE * my, int * params )
+static int math_atan( INSTANCE * my, intptr_t * params )
 {
     return atan(( double )( params[0] * 180000.0 / M_PI )) * 1000 ;
 }
 
 /* --------------------------------------------------------------------------- */
 
-static int math_atan2( INSTANCE * my, int * params )
+static int math_atan2( INSTANCE * my, intptr_t * params )
 {
     return atan2(( double )( params[0] * 180000.0 / M_PI ), ( double )( params[1] * 180000.0 / M_PI )) * 1000 ;
 }
 
 /* --------------------------------------------------------------------------- */
 
-static int math_isinf( INSTANCE * my, int * params )
+static int math_isinf( INSTANCE * my, intptr_t * params )
 {
     double param = ( double ) *( float * ) & params[0] ;
     return isinf( param );
@@ -121,7 +121,7 @@ static int math_isinf( INSTANCE * my, int * params )
 
 /* --------------------------------------------------------------------------- */
 
-static int math_isnan( INSTANCE * my, int * params )
+static int math_isnan( INSTANCE * my, intptr_t * params )
 {
     double param = ( double ) *( float * ) & params[0] ;
     return isnan( param );
@@ -129,7 +129,7 @@ static int math_isnan( INSTANCE * my, int * params )
 
 /* --------------------------------------------------------------------------- */
 
-static int math_finite( INSTANCE * my, int * params )
+static int math_finite( INSTANCE * my, intptr_t * params )
 {
     double param = ( double ) *( float * ) & params[0] ;
     return finite ( param );
@@ -137,7 +137,7 @@ static int math_finite( INSTANCE * my, int * params )
 
 /* --------------------------------------------------------------------------- */
 
-static int math_fget_angle( INSTANCE * my, int * params )
+static int math_fget_angle( INSTANCE * my, intptr_t * params )
 {
     double dx = params[2] - params[0] ;
     double dy = params[3] - params[1] ;
@@ -152,7 +152,7 @@ static int math_fget_angle( INSTANCE * my, int * params )
 
 /* --------------------------------------------------------------------------- */
 
-static int math_fget_dist( INSTANCE * my, int * params )
+static int math_fget_dist( INSTANCE * my, intptr_t * params )
 {
     double dx = params[2] - params[0] ;
     double dy = params[3] - params[1] ;
@@ -162,7 +162,7 @@ static int math_fget_dist( INSTANCE * my, int * params )
 
 /* --------------------------------------------------------------------------- */
 
-static int math_near_angle( INSTANCE * my, int * params )
+static int math_near_angle( INSTANCE * my, intptr_t * params )
 {
     int angle = params[0] ;
     int dest  = params[1] ;
@@ -192,7 +192,7 @@ static int math_near_angle( INSTANCE * my, int * params )
 
 /* --------------------------------------------------------------------------- */
 
-static int math_get_distx( INSTANCE * my, int * params )
+static int math_get_distx( INSTANCE * my, intptr_t * params )
 {
     double angle = params[0] * M_PI / 180000.0 ;
     return ( int )( params[1] * cos( angle ) ) ;
@@ -200,7 +200,7 @@ static int math_get_distx( INSTANCE * my, int * params )
 
 /* --------------------------------------------------------------------------- */
 
-static int math_get_disty( INSTANCE * my, int * params )
+static int math_get_disty( INSTANCE * my, intptr_t * params )
 {
     double angle = params[0] * M_PI / 180000.0 ;
     return ( int )( params[1] * -sin( angle ) ) ;
